Function reference

Function reference

Top-level

JuliaInterpreter.@interpret — Macro.

@interpret f(args; kwargs...)

Evaluate f on the specified arguments using the interpreter.

Example

julia> a = [1, 7]
2-element Array{Int64,1}:
 1
 7

julia> sum(a)
8

julia> @interpret sum(a)
8

source

Frame creation

JuliaInterpreter.enter_call — Function.

frame = enter_call(f, args...; kwargs...)

Build a JuliaStackFrame ready to execute f with the specified positional and keyword arguments.

Example

julia> mymethod(x) = x+1
mymethod (generic function with 1 method)

julia> JuliaInterpreter.enter_call(mymethod, 1)
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x) in Main at none:1, CodeInfo(
1 ─ %1 = ($(QuoteNode(+)))(x, 1)
└──      return %1
), Union{Compiled, TypeMapEntry}[#undef, #undef], JuliaInterpreter.BreakpointState[#undef, #undef], BitSet([1]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some(1)], Any[#undef, #undef], Any[], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict(Symbol("#self#")=>1,:x=>2), Any[])

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 2 methods)

julia> JuliaInterpreter.enter_call(mymethod, [1.0, 2.0])
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Union{Compiled, TypeMapEntry}[#undef], JuliaInterpreter.BreakpointState[#undef], BitSet([]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some([1.0, 2.0])], Any[#undef], Any[Float64], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict(Symbol("#self#")=>1,:x=>2), Any[])

For a @generated function you can use enter_call((f, true), args...; kwargs...) to execute the generator of a @generated function, rather than the code that would be created by the generator.

See enter_call_expr for a similar approach based on expressions.

source

JuliaInterpreter.enter_call_expr — Function.

frame = enter_call_expr(expr; enter_generated=false)

Build a JuliaStackFrame ready to execute the expression expr. Set enter_generated=true if you want to execute the generator of a @generated function, rather than the code that would be created by the generator.

Example

julia> mymethod(x) = x+1
mymethod (generic function with 1 method)

julia> JuliaInterpreter.enter_call_expr(:($mymethod(1)))
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x) in Main at none:1, CodeInfo(
1 ─ %1 = ($(QuoteNode(+)))(x, 1)
└──      return %1
), Union{Compiled, TypeMapEntry}[#undef, #undef], JuliaInterpreter.BreakpointState[#undef, #undef], BitSet([1]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some(1)], Any[#undef, #undef], Any[], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict(Symbol("#self#")=>1,:x=>2), Any[])

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 2 methods)

julia> a = [1.0, 2.0]
2-element Array{Float64,1}:
 1.0
 2.0

julia> JuliaInterpreter.enter_call_expr(:($mymethod($a)))
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Union{Compiled, TypeMapEntry}[#undef], JuliaInterpreter.BreakpointState[#undef], BitSet([]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some([1.0, 2.0])], Any[#undef], Any[Float64], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict(Symbol("#self#")=>1,:x=>2), Any[])

See enter_call for a similar approach not based on expressions.

source

JuliaInterpreter.build_frame — Function.

frame = build_frame(framecode::JuliaFrameCode, frameargs, lenv)

Construct a new JuliaStackFrame for framecode, given lowered-code arguments frameargs and static parameters lenv. See JuliaInterpreter.prepare_call for information about how to prepare the inputs.

source

JuliaInterpreter.determine_method_for_expr — Function.

framecode, frameargs, lenv, argtypes = determine_method_for_expr(expr; enter_generated = false)

Prepare all the information needed to execute a particular :call expression expr. For example, try JuliaInterpreter.determine_method_for_expr(:(sum([1,2]))). See JuliaInterpreter.prepare_call for information about the outputs.

source

JuliaInterpreter.prepare_args — Function.

frun, allargs = prepare_args(fcall, fargs, kwargs)

Prepare the complete argument sequence for a call to fcall. fargs = [fcall, args...] is a list containing both fcall (the #self# slot in lowered code) and the positional arguments supplied to fcall. kwargs is a list of keyword arguments, supplied either as list of expressions :(kwname=kwval) or pairs :kwname=>kwval.

For non-keyword methods, frun === fcall, but for methods with keywords frun will be the keyword-sorter function for fcall.

Example

julia> mymethod(x) = 1
mymethod (generic function with 1 method)

julia> mymethod(x, y; verbose=false) = nothing
mymethod (generic function with 2 methods)

julia> JuliaInterpreter.prepare_args(mymethod, [mymethod, 15], ())
(mymethod, Any[mymethod, 15])

julia> JuliaInterpreter.prepare_args(mymethod, [mymethod, 1, 2], [:verbose=>true])
(getfield( Symbol("#kw##mymethod"))(), Any[#kw##mymethod(), (verbose = true,), mymethod, 1, 2])

source

JuliaInterpreter.prepare_call — Function.

framecode, frameargs, lenv, argtypes = prepare_call(f, allargs; enter_generated=false)

Prepare all the information needed to execute lowered code for f given arguments allargs. f and allargs are the outputs of prepare_args. For @generated methods, set enter_generated=true if you want to extract the lowered code of the generator itself.

On return framecode is the JuliaFrameCode of the method. frameargs contains the actual arguments needed for executing this frame (for generators, this will be the types of allargs); lenv is the "environment", i.e., the static parameters for f given allargs. argtypes is the Tuple-type for this specific call (equivalent to the signature of the MethodInstance).

Example

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 1 method)

julia> framecode, frameargs, lenv, argtypes = JuliaInterpreter.prepare_call(mymethod, [mymethod, [1.0,2.0]]);

julia> framecode
JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Union{Compiled, TypeMapEntry}[#undef], JuliaInterpreter.BreakpointState[#undef], BitSet([]), false, false, true)

julia> frameargs
2-element Array{Any,1}:
 mymethod
 [1.0, 2.0]

julia> lenv
svec(Float64)

julia> argtypes
Tuple{typeof(mymethod),Array{Float64,1}}

source

JuliaInterpreter.prepare_thunk — Function.

frame = prepare_thunk(mod::Module, expr::Expr)

Prepare expr for evaluation in mod. expr should be a "straightforward" expression, one that does not require special top-level handling (see JuliaInterpreter.split_expressions).

source

JuliaInterpreter.split_expressions — Function.

modexs, docexprs = split_expressions(mod::Module, expr::Expr; extract_docexprs=false)

Break expr into a list modexs of sequential blocks. This is often needed when expr needs to be evaluated at top level.

modexs[i] is a (mod::Module, ex::Expr) tuple, where ex is to be evaluated in mod.

Toplevel evaluation

For code that defines new structs, new methods, or new macros, it can be important to evaluate these expressions carefully:

stack = JuliaStackFrame[]
for modex in modexs    # or use `for (mod, ex) in modexs` to split the tuple
    frame = JuliaInterpreter.prepare_thunk(modex)
    while true
        JuliaInterpreter.through_methoddef_or_done!(stack, frame) === nothing && break
    end
end

The while loop here deserves some explanation. Occasionally, a frame may define new methods (e.g., anonymous or local functions) and then call those methods. In such cases, running the entire frame as a single block (e.g., with JuliaInterpreter.finish_and_return! can trigger "method is too new..." errors. Instead, the approach above runs each frame, but returns to the caller after any new method is defined. When this loop is running at top level (e.g., in the REPL), this allows the world age to update and thus avoid "method is too new..." errors.

Putting the above nested loop inside a function defeats its purpose, because inside a compiled function the world age will not update. If necessary, use the following strategy:

Core.eval(somemodule, Expr(:toplevel, quote
    body
))

where body contains the nested loop, plus any preparatory statements required to make the necessary variables available at top level in somemodule.

source

JuliaInterpreter.get_call_framecode — Function.

framecode, lenv = get_call_framecode(fargs, parentframe::JuliaFrameCode, idx::Int)

Return the framecode and environment for a call specified by fargs = [f, args...] (see prepare_args). parentframecode is the caller, and idx is the program-counter index. If possible, framecode will be looked up from the local method tables of parentframe.

source

JuliaInterpreter.optimize! — Function.

optimize!(code::CodeInfo, mod::Module)

Perform minor optimizations on the lowered AST in code to reduce execution time of the interpreter. Currently it looks up GlobalRefs (for which it needs mod to know the scope in which this will run) and ensures that no statement includes nested :call expressions (splitting them out into multiple SSA-form statements if needed).

source

Frame execution

JuliaInterpreter.Compiled — Type.

Compiled is a trait indicating that any :call expressions should be evaluated using Julia's normal compiled-code evaluation. The alternative is to pass stack=JuliaStackFrame[], which will cause all calls to be evaluated via the interpreter.

source

JuliaInterpreter.step_expr! — Function.

pc = step_expr!(stack, frame)

Execute the next statement in frame. pc is the new program counter, or nothing if execution terminates. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

source

JuliaInterpreter.finish! — Function.

pc = finish!(stack, frame, pc=frame.pc[])

Run frame until execution terminates. pc is the program counter for the final statement. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

If execution hits a breakpoint, then pc is a reference to the breakpoint. stack[end], if not running in Compiled() mode, will contain the frame in which the breakpoint was hit.

source

JuliaInterpreter.finish_and_return! — Function.

ret = finish_and_return!(stack, frame, istoplevel::Bool=false)
ret = finish_and_return!(stack, frame, pc, istoplevel::Bool)

Run frame until execution terminates, and pass back the computed return value. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

If execution hits a breakpoint, then ret is a reference to the breakpoint. stack[end], if not running in Compiled() mode, will contain the frame in which the breakpoint was hit.

Optionally supply the starting pc, if you don't want to start at the current location in frame.

source

JuliaInterpreter.finish_stack! — Function.

ret = finish_stack!(stack)

Completely unwind stack, finishing it frame-by-frame. If execution hits a breakpoint, ret will be a reference to the breakpoint.

source

JuliaInterpreter.get_return — Function.

ret = get_return(frame, pc=frame.pc[])

Get the return value of frame. Throws an error if pc does not point to a return expression. frame must have already been executed so that the return value has been computed (see, e.g., JuliaInterpreter.finish!).

source

JuliaInterpreter.next_until! — Function.

next_until!(predicate, stack, frame, pc=frame.pc[])

Step through statements of frame until the next statement satifies predicate(stmt).

source

JuliaInterpreter.through_methoddef_or_done! — Function.

through_methoddef_or_done!(stack, frame)

Runs frame at top level until it either finishes (e.g., hits a return statement) or defines a new method.

source

JuliaInterpreter.evaluate_call! — Function.

ret = evaluate_call!(Compiled(), frame::JuliaStackFrame, call_expr, pc)
ret = evaluate_call!(stack,      frame::JuliaStackFrame, call_expr, pc)

Evaluate a :call expression call_expr in the context of frame. The first causes it to be executed using Julia's normal dispatch (compiled code), whereas the second recurses in via the interpreter. stack should be a vector of JuliaStackFrame.

source

JuliaInterpreter.evaluate_foreigncall! — Function.

ret = evaluate_foreigncall!(stack, frame::JuliaStackFrame, call_expr, pc)

Evaluate a :foreigncall (from a ccall) statement callexpr in the context of frame. stack and pc are unused, but supplied for consistency with evaluate_call!.

source

JuliaInterpreter.maybe_evaluate_builtin — Function.

ret = maybe_evaluate_builtin(frame, call_expr)

If call_expr is to a builtin function, evaluate it, returning the result inside a Some wrapper. Otherwise, return call_expr.

source

Breakpoints

JuliaInterpreter.Breakpoints.@breakpoint — Macro.

@breakpoint f(args...) condition=nothing
@breakpoint f(args...) line condition=nothing

Break upon entry, or at the specified line number, in the method called by f(args...). Optionally supply a condition expressed in terms of the arguments and internal variables of the method. If line is supplied, it must be a literal integer.

Example

Suppose a method mysum is defined as follows, where the numbers to the left are the line number in the file:

12 function mysum(A)
13     s = zero(eltype(A))
14     for a in A
15         s += a
16     end
17     return s
18 end

Then

@breakpoint mysum(A) 15 s>10

would cause execution of the loop to break whenever s>10.

source

JuliaInterpreter.Breakpoints.breakpoint — Function.

breakpoint(f, sig)
breakpoint(f, sig, line)
breakpoint(f, sig, condition)
breakpoint(f, sig, line, condition)
breakpoint(...; enter_generated=false)

Add a breakpoint to f with the specified argument types sig. Optionally specify an absolute line number line in the source file; the default is to break upon entry at the first line of the body. Without condition, the breakpoint will be triggered every time it is encountered; the second only if condition evaluates to true. condition should be written in terms of the arguments and local variables of f.

Example

function radius2(x, y)
    return x^2 + y^2
end

breakpoint(radius2, Tuple{Int,Int}, :(y > x))

source

breakpoint(method::Method)
breakpoint(method::Method, line)
breakpoint(method::Method, condition::Expr)
breakpoint(method::Method, line, condition::Expr)

Add a breakpoint to method.

source

breakpoint(f)
breakpoint(f, condition)

Break-on-entry to all methods of f.

source

JuliaInterpreter.Breakpoints.enable — Function.

enable(bp::BreakpointRef)

Enable breakpoint bp.

source

enable()

Enable all breakpoints.

source

JuliaInterpreter.Breakpoints.disable — Function.

disable(bp::BreakpointRef)

Disable breakpoint bp. Disabled breakpoints can be re-enabled with enable.

source

disable()

Disable all breakpoints.

source

JuliaInterpreter.Breakpoints.remove — Function.

remove(bp::BreakpointRef)

Remove (delete) breakpoint bp. Removed breakpoints cannot be re-enabled.

source

remove()

Remove all breakpoints.

source

Types

JuliaInterpreter.JuliaStackFrame — Type.

JuliaStackFrame represents the current execution state in a particular call frame.

Important fields:

code: the JuliaFrameCode for this frame

locals: a vector containing the input arguments and named local variables for this frame. The indexing corresponds to the names in frame.code.code.slotnames.

ssavalues: a vector containing the Static Single Assignment values produced at the current state of execution

sparams: the static type parameters, e.g., for f(x::Vector{T}) where T this would store the value of T given the particular input x.

pc: the JuliaProgramCounter that typically represents the current position during execution. However, note that some internal functions instead maintain the pc as a local variable, and only update the frame's pc when pushing a frame on the stack.

source

JuliaInterpreter.JuliaFrameCode — Type.

JuliaFrameCode holds static information about a method or toplevel code. One JuliaFrameCode can be shared by many JuliaStackFrame calling frames.

Important fields:

scope: the Method or Module in which this frame is to be evaluated

code: the CodeInfo object storing (optimized) lowered code

methodtables: a vector, each entry potentially stores a "local method table" for the corresponding :call expression in code (undefined entries correspond to statements that do not contain :call expressions)

used: a BitSet storing the list of SSAValues that get referenced by later statements.

source

JuliaInterpreter.JuliaProgramCounter — Type.

JuliaProgramCounter(next_stmt::Int)

A wrapper specifying the index of the next statement in the lowered code to be executed.

source

JuliaInterpreter.BreakpointState — Type.

BreakpointState(isactive=true, condition=JuliaInterpreter.truecondition)

BreakpointState represents a breakpoint at a particular statement in a JuliaFrameCode. isactive indicates whether the breakpoint is currently enabled or disabled. condition is a function that accepts a single JuliaStackFrame, and condition(frame) must return either true or false. Execution will stop at a breakpoint only if isactive and condition(frame) both evaluate as true. The default condition always returns true.

To create these objects, see breakpoint.

source

JuliaInterpreter.Breakpoints.BreakpointRef — Type.

BreakpointRef(framecode, stmtidx)
BreakpointRef(framecode, stmtidx, err)

A reference to a breakpoint at a particular statement index stmtidx in framecode. If the break was due to an error, supply that as well.

source

Internal storage

JuliaInterpreter.framedict — Constant.

framedict[method] returns the JuliaFrameCode for method. For @generated methods, see genframedict.

source

JuliaInterpreter.genframedict — Constant.

genframedict[(method,argtypes)] returns the JuliaFrameCode for a @generated method method, for the particular argument types argtypes.

The framecodes stored in genframedict are for the code returned by the generator (i.e, what will run when you call the method on particular argument types); for the generator itself, its framecode would be stored in framedict.

source

JuliaInterpreter.compiled_methods — Constant.

meth ∈ compiled_methods indicates that meth should be run using Compiled rather than recursed into via the interpreter.

source

Utilities

JuliaInterpreter.@lookup — Macro.

rhs = @lookup(frame, node)
rhs = @lookup(mod, frame, node)

This macro looks up previously-computed values referenced as SSAValues, SlotNumbers, GlobalRefs, QuoteNode, sparam or exception reference expression. It will also lookup symbols in moduleof(frame); this can be supplied ahead-of-time via the 3-argument version. If none of the above apply, the value of node will be returned.

source

JuliaInterpreter.iswrappercall — Function.

Determine whether we are calling a function for which the current function is a wrapper (either because of optional arguments or becaue of keyword arguments).

source

JuliaInterpreter.isdocexpr — Function.

isdocexpr(ex)

Test whether expression ex is a @doc expression.

source

JuliaInterpreter.isglobalref — Function.

isglobalref(g, mod, name)

Tests whether g is equal to GlobalRef(mod, name).

source

JuliaInterpreter.statementnumber — Function.

stmtidx = statementnumber(frame, line)

Return the index of the first statement in frame's CodeInfo that corresponds to line.

source

framecode, stmtidx = statementnumber(method, line)

Return the index of the first statement in framecode that corresponds to the given line in method.

source

JuliaInterpreter.Variable — Type.

Variable is a struct representing a variable with an asigned value. By calling the function locals[@ref] on a JuliaStackFrame[@ref] a Vector of Variable's is returned.

Important fields:

value::Any: the value of the local variable

name::Symbol: the name of the variable as given in the source code

isparam::Bool: if the variable is a type parameter, for example T in f(x::T) where {T} = x .

source

JuliaInterpreter.locals — Function.

local_variables = locals(frame::JuliaStackFrame)::Vector{Variable}

Return the local variables as a vector of Variable[@ref].

source

CodeTracking.whereis — Function.

loc = whereis(frame, pc=frame.pc[])

Return the file and line number for frame at pc. If this cannot be determined, loc == nothing. Otherwise loc == (filepath, line).

When frame represents top-level code,

source