Function reference

Function reference

Top-level

JuliaInterpreter.@interpret — Macro.

@interpret f(args; kwargs...)

Evaluate f on the specified arguments using the interpreter.

Example

julia> a = [1, 7]
2-element Array{Int64,1}:
 1
 7

julia> sum(a)
8

julia> @interpret sum(a)
8

source

Frame creation

JuliaInterpreter.enter_call — Function.

frame = enter_call(f, args...; kwargs...)

Build a JuliaStackFrame ready to execute f with the specified positional and keyword arguments.

Example

julia> mymethod(x) = x+1
mymethod (generic function with 1 method)

julia> JuliaInterpreter.enter_call(mymethod, 1)
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x) in Main at none:1, CodeInfo(
1 ─ %1 = ($(QuoteNode(+)))(x, 1)
└──      return %1
), Core.TypeMapEntry[#undef, #undef], BitSet([1]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some(1)], Any[#undef, #undef], Any[], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict{Symbol,Int64}(), Any[])

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 2 methods)

julia> JuliaInterpreter.enter_call(mymethod, [1.0, 2.0])
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Core.TypeMapEntry[#undef], BitSet([]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some([1.0, 2.0])], Any[#undef], Any[Float64], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict{Symbol,Int64}(), Any[])

For a @generated function you can use enter_call((f, true), args...; kwargs...) to execute the generator of a @generated function, rather than the code that would be created by the generator.

See enter_call_expr for a similar approach based on expressions.

source

JuliaInterpreter.enter_call_expr — Function.

frame = enter_call_expr(expr; enter_generated=false)

Build a JuliaStackFrame ready to execute the expression expr. Set enter_generated=true if you want to execute the generator of a @generated function, rather than the code that would be created by the generator.

Example

julia> mymethod(x) = x+1
mymethod (generic function with 1 method)

julia> JuliaInterpreter.enter_call_expr(:($mymethod(1)))
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x) in Main at none:1, CodeInfo(
1 ─ %1 = ($(QuoteNode(+)))(x, 1)
└──      return %1
), Core.TypeMapEntry[#undef, #undef], BitSet([1]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some(1)], Any[#undef, #undef], Any[], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict{Symbol,Int64}(), Any[])

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 2 methods)

julia> a = [1.0, 2.0]
2-element Array{Float64,1}:
 1.0
 2.0

julia> JuliaInterpreter.enter_call_expr(:($mymethod($a)))
JuliaStackFrame(JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Core.TypeMapEntry[#undef], BitSet([]), false, false, true), Union{Nothing, Some{Any}}[Some(mymethod), Some([1.0, 2.0])], Any[#undef], Any[Float64], Int64[], Base.RefValue{Any}(nothing), Base.RefValue{JuliaInterpreter.JuliaProgramCounter}(JuliaProgramCounter(1)), Dict{Symbol,Int64}(), Any[])

See enter_call for a similar approach not based on expressions.

source

JuliaInterpreter.build_frame — Function.

frame = build_frame(framecode::JuliaFrameCode, frameargs, lenv)

Construct a new JuliaStackFrame for framecode, given lowered-code arguments frameargs and static parameters lenv. See JuliaInterpreter.prepare_call for information about how to prepare the inputs.

source

JuliaInterpreter.determine_method_for_expr — Function.

framecode, frameargs, lenv, argtypes = determine_method_for_expr(expr; enter_generated = false)

Prepare all the information needed to execute a particular :call expression expr. For example, try JuliaInterpreter.determine_method_for_expr(:(sum([1,2]))). See JuliaInterpreter.prepare_call for information about the outputs.

source

JuliaInterpreter.prepare_args — Function.

frun, allargs = prepare_args(fcall, fargs, kwargs)

Prepare the complete argument sequence for a call to fcall. fargs = [fcall, args...] is a list containing both fcall (the #self# slot in lowered code) and the positional arguments supplied to fcall. kwargs is a list of keyword arguments, supplied either as list of expressions :(kwname=kwval) or pairs :kwname=>kwval.

For non-keyword methods, frun === fcall, but for methods with keywords frun will be the keyword-sorter function for fcall.

Example

julia> mymethod(x) = 1
mymethod (generic function with 1 method)

julia> mymethod(x, y; verbose=false) = nothing
mymethod (generic function with 2 methods)

julia> JuliaInterpreter.prepare_args(mymethod, [mymethod, 15], ())
(mymethod, Any[mymethod, 15])

julia> JuliaInterpreter.prepare_args(mymethod, [mymethod, 1, 2], [:verbose=>true])
(getfield( Symbol("#kw##mymethod"))(), Any[#kw##mymethod(), (verbose = true,), mymethod, 1, 2])

source

JuliaInterpreter.prepare_call — Function.

framecode, frameargs, lenv, argtypes = prepare_call(f, allargs; enter_generated=false)

Prepare all the information needed to execute lowered code for f given arguments allargs. f and allargs are the outputs of prepare_args. For @generated methods, set enter_generated=true if you want to extract the lowered code of the generator itself.

On return framecode is the JuliaFrameCode of the method. frameargs contains the actual arguments needed for executing this frame (for generators, this will be the types of allargs); lenv is the "environment", i.e., the static parameters for f given allargs. argtypes is the Tuple-type for this specific call (equivalent to the signature of the MethodInstance).

Example

julia> mymethod(x::Vector{T}) where T = 1
mymethod (generic function with 1 method)

julia> framecode, frameargs, lenv, argtypes = JuliaInterpreter.prepare_call(mymethod, [mymethod, [1.0,2.0]]);

julia> framecode
JuliaInterpreter.JuliaFrameCode(mymethod(x::Array{T,1}) where T in Main at none:1, CodeInfo(
1 ─     return 1
), Core.TypeMapEntry[#undef], BitSet([]), false, false, true)

julia> frameargs
2-element Array{Any,1}:
 mymethod
 [1.0, 2.0]

julia> lenv
svec(Float64)

julia> argtypes
Tuple{typeof(mymethod),Array{Float64,1}}

source

JuliaInterpreter.get_call_framecode — Function.

framecode, lenv = get_call_framecode(fargs, parentframe::JuliaFrameCode, idx::Int)

Return the framecode and environment for a call specified by fargs = [f, args...] (see prepare_args). parentframecode is the caller, and idx is the program-counter index. If possible, framecode will be looked up from the local method tables of parentframe.

source

JuliaInterpreter.optimize! — Function.

optimize!(code::CodeInfo, mod::Module)

Perform minor optimizations on the lowered AST in code to reduce execution time of the interpreter. Currently it looks up GlobalRefs (for which it needs mod to know the scope in which this will run) and ensures that no statement includes nested :call expressions (splitting them out into multiple SSA-form statements if needed).

source

Frame execution

JuliaInterpreter.Compiled — Type.

Compiled is a trait indicating that any :call expressions should be evaluated using Julia's normal compiled-code evaluation. The alternative is to pass stack=JuliaStackFrame[], which will cause all calls to be evaluated via the interpreter.

source

JuliaInterpreter.step_expr! — Function.

pc = step_expr!(stack, frame)

Execute the next statement in frame. pc is the new program counter, or nothing if execution terminates. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

source

JuliaInterpreter.finish! — Function.

pc = finish!(stack, frame, pc=frame.pc[])

Run frame until execution terminates. pc is the program counter for the final statement. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

source

JuliaInterpreter.finish_and_return! — Function.

ret = finish_and_return!(stack, frame, pc=frame.pc[])

Run frame until execution terminates, and pass back the computed return value. stack controls call evaluation; stack = Compiled() evaluates :call expressions by normal dispatch, whereas a vector of JuliaStackFrames will use recursive interpretation.

source

JuliaInterpreter.next_until! — Function.

next_until!(predicate, stack, frame, pc=frame.pc[])

Step through statements of frame until the next statement satifies predicate(stmt).

source

JuliaInterpreter.evaluate_call! — Function.

ret = evaluate_call!(Compiled(), frame::JuliaStackFrame, call_expr, pc)
ret = evaluate_call!(stack,      frame::JuliaStackFrame, call_expr, pc)

Evaluate a :call expression call_expr in the context of frame. The first causes it to be executed using Julia's normal dispatch (compiled code), whereas the second recurses in via the interpreter. stack should be a vector of JuliaStackFrame.

source

JuliaInterpreter.evaluate_foreigncall! — Function.

ret = evaluate_foreigncall!(stack, frame::JuliaStackFrame, call_expr, pc)

Evaluate a :foreigncall (from a ccall) statement callexpr in the context of frame. stack and pc are unused, but supplied for consistency with evaluate_call!.

source

JuliaInterpreter.maybe_evaluate_builtin — Function.

ret = maybe_evaluate_builtin(frame, call_expr)

If call_expr is to a builtin function, evaluate it, returning the result inside a Some wrapper. Otherwise, return call_expr.

source

Types

JuliaInterpreter.JuliaStackFrame — Type.

JuliaStackFrame represents the current execution state in a particular call frame.

Important fields:

code: the JuliaFrameCode for this frame

locals: a vector containing the input arguments and named local variables for this frame. The indexing corresponds to the names in frame.code.code.slotnames.

ssavalues: a vector containing the Static Single Assignment values produced at the current state of execution

sparams: the static type parameters, e.g., for f(x::Vector{T}) where T this would store the value of T given the particular input x.

pc: the JuliaProgramCounter that typically represents the current position during execution. However, note that some internal functions instead maintain the pc as a local variable, and only update the frame's pc when pushing a frame on the stack.

source

JuliaInterpreter.JuliaFrameCode — Type.

JuliaFrameCode holds static information about a method or toplevel code. One JuliaFrameCode can be shared by many JuliaFrameState calling frames.

Important fields:

scope: the Method or Module in which this frame is to be evaluated

code: the CodeInfo object storing (optimized) lowered code

methodtables: a vector, each entry potentially stores a "local method table" for the corresponding :call expression in code (undefined entries correspond to statements that do not contain :call expressions)

used: a BitSet storing the list of SSAValues that get referenced by later statements.

source

JuliaInterpreter.JuliaProgramCounter — Type.

JuliaProgramCounter(next_stmt::Int)

A wrapper specifying the index of the next statement in the lowered code to be executed.

source

Internal storage

JuliaInterpreter.framedict — Constant.

framedict[method] returns the JuliaFrameCode for method. For @generated methods, see genframedict.

source

JuliaInterpreter.genframedict — Constant.

genframedict[(method,argtypes)] returns the JuliaFrameCode for a @generated method method, for the particular argument types argtypes.

The framecodes stored in genframedict are for the code returned by the generator (i.e, what will run when you call the method on particular argument types); for the generator itself, its framecode would be stored in framedict.

source

Utilities

JuliaInterpreter.iswrappercall — Function.

Determine whether we are calling a function for which the current function is a wrapper (either because of optional arguments or becaue of keyword arguments).

source